requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
black==24.8.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==45.0.2
distlib==0.3.9
dnspython==2.6.1
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.16.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy==1.14.1
mypy-extensions==1.1.0
nodeenv==1.9.1
orjson==3.10.15
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.4
pydantic-settings==2.8.1
pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.44.0
tomli==2.2.1
typer==0.15.4
typing-extensions==4.13.2
ua-parser==0.18.0
ujson==5.10.0
user-agents==2.2.0
uvicorn==0.33.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==0.24.0
websockets==13.1
